requirements.txt
asgiref==3.8.1
Django==5.2
geographiclib==2.0
geopy==2.4.1
sqlparse==0.5.3
django-phonenumber-field[phonenumbers]==7.0.0
Pillow
whitenoise
bs4
requests
mysql
telegram
python-telegram-bot
python-telegram-bot[webhook]
django-allauth==64.0.0
PyJWT>=2.10
psycopg2-binary>=2.9
gunicorn>=21.2
